% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_app.R
\name{create_dash}
\alias{create_dash}
\title{Create distill themed dashboard}
\usage{
create_dash(path, title = NULL)
}
\arguments{
\item{path}{Path to the new directory}

\item{title}{App title}
}
\description{
Create distill themed dashboard
}
