% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_app.R
\name{create_simple_app}
\alias{create_simple_app}
\title{Create distill themed app}
\usage{
create_simple_app(path = ".", title = NULL)
}
\arguments{
\item{path}{Path to the new directory}

\item{title}{App title}
}
\description{
Create distill themed app
}
