% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_app.R
\name{criar_dash_r6}
\alias{criar_dash_r6}
\title{Create distill themed app}
\usage{
criar_dash_r6(path = ".", title = NULL)
}
\arguments{
\item{path}{Path to the new directory}

\item{title}{App title}
}
\description{
Create distill themed app
}
